# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messenger backend api",
        "description": "Messenger backend",
        "contact": {
            "email": "tcom.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://crypt-talk.test/api",
            "description": "Dev server"
        },
        {
            "url": "http://95.217.156.195/api",
            "description": "Dev server"
        }
    ],
    "paths": {
...
        "/conversation/index": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Get all conversations for authenticated user",
                "operationId": "9281aa17648b606969b830d642019834",
                "responses": {
                    "200": {
                        "description": "All user conversations"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/message/conversation/{friend_id}": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Get conversation with a friend",
                "operationId": "14d346ddc59e298197daeb51531d6e33",
                "parameters": [
                    {
                        "name": "friend_id",
                        "in": "path",
                        "description": "ID of the friend to fetch conversation with",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/conversation/send-message": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Send message",
                "operationId": "b667afe5855d13afe01cbd42da75eb94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "conversationId",
                                    "isEncrypted"
                                ],
                                "properties": {
                                    "conversationId": {
                                        "type": "integer",
                                        "default": 666
                                    },
                                    "isEncrypted": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "text": {
                                        "type": "text",
                                        "default": "hey whatsupp"
                                    },
                                    "ai": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/conversation/send-message-ai": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Send message to AI",
                "operationId": "df232dd9b167cad2b908835568ec3129",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "conversationId",
                                    "isEncrypted"
                                ],
                                "properties": {
                                    "conversationId": {
                                        "type": "integer",
                                        "default": 4
                                    },
                                    "isEncrypted": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "text": {
                                        "type": "text",
                                        "default": "hey whatsupp"
                                    },
                                    "ai": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorize with bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}